Fixed pinned tab navigation issue
Tab navigation is re-enabled if the active tab is a pinned tab (even if
the user has selected to skip pinned tabs). So you can always navigate from a
pinned tab but if you want to navigate to a pinned tab you should first
deselect the option to skip pinned tabs.

Signed-off-by: Mehmet Baker <mehmet.baker@gmail.com>